requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
requests==2.31.0
numpy~=2.0.1
pillow~=10.4.0